Refactor the wazo_load_cli for installation
