+++ tox.ini
@@ -21,7 +21,7 @@ commands = pre-commit run --all-files
 [testenv:cli]
 changedir = wazo_load_cli
 commands =
-    pytest -s -v tests/
+    pytest -s -v wlctl/tests/
 deps =
     -rwazo_load_cli/requirements.txt
     -rwazo_load_cli/test-requirements.txt
