+++ wazo_load_cli/wlctl/modules/templates/baresip_job.tpl
@@ -1,0 +1,7 @@
+- node:
+  cmd: {{ CMD }}
+  env:
+    DEBUG: {{ DEBUG }}
+    LINE: {{ LINE }}
+    CALL_DURATION: {{ CALL_DURATION }}
+    STACK: {{ STACK }}
